refactored files and structure
