+++ app/src/main/AndroidManifest.xml
@@ -73,29 +73,17 @@
             android:name=".food.food_notif_service"
             android:enabled="true"
             android:exported="true" />
-        <service
-            android:name=".food.food_sms_service"
-            android:enabled="true"
-            android:exported="true" />
-        <service
-            android:name=".exercise.exercise_sms_service"
-            android:enabled="true"
-            android:exported="true" />
-        <service
-            android:name=".food.food_notif_service"
-            android:enabled="true"
-            android:exported="true" />
 
         <receiver
-            android:name=".utilities.NotificationPublisher"
+            android:name=".publishers.NotificationPublisher"
             android:enabled="true"
             android:exported="true" />
         <receiver
-            android:name=".food.FoodSMSpublisher"
+            android:name=".publishers.FoodSMSpublisher"
             android:enabled="true"
             android:exported="true" />
         <receiver
-            android:name=".exercise.ExerciseSMSpublisher"
+            android:name=".publishers.ExerciseSMSpublisher"
             android:enabled="true"
             android:exported="true"></receiver>
     </application>
