--- app/src/main/java/edu/bu/fitnessfriend/fitnessfriend/exercise/ExerciseSMSpublisher.java
@@ -1,13 +1,0 @@
-package edu.bu.fitnessfriend.fitnessfriend.exercise;
-
-import android.content.BroadcastReceiver;
-import android.content.Context;
-import android.content.Intent;
-
-public class ExerciseSMSpublisher extends BroadcastReceiver {
-
-    @Override
-    public void onReceive(Context context, Intent intent) {
-        exercise_sms_service.sendSmsMessage(context);
-    }
-}
