--- app/src/main/java/edu/bu/fitnessfriend/fitnessfriend/utilities/NotificationPublisher.java
@@ -1,20 +1,0 @@
-package edu.bu.fitnessfriend.fitnessfriend.utilities;
-
-import android.app.Notification;
-import android.app.NotificationManager;
-import android.content.BroadcastReceiver;
-import android.content.Context;
-import android.content.Intent;
-
-public class NotificationPublisher extends BroadcastReceiver {
-
-    @Override
-    public void onReceive(Context context, Intent intent) {
-        NotificationManager notificationManager = (NotificationManager)
-                context.getSystemService(Context.NOTIFICATION_SERVICE);
-
-        Notification notification = intent.getParcelableExtra("notification");
-        int notificationID = intent.getIntExtra("notification_id",0);
-        notificationManager.notify(notificationID,notification);
-    }
-}
